// src/app/modules/student/student.model.ts
import { Schema, model } from 'mongoose';
import { Guardian, UserName, LocalGurdian, Student } from './student.interface';

const UserNameScema = new Schema<UserName>({
  firstName: { type: String, required: true },
  middleName: { type: String },
  lastName: { type: String, required: true },
});

const GuardianScema = new Schema<Guardian>({
  fatherName: { type: String, required: true },
  fatherOccupation: { type: String, required: true },
  fatherContactNo: { type: String, required: true },
  motherName: { type: String, required: true },
  motherOccupation: { type: String, required: true },
  motherContactNo: { type: String, required: true },
});

const LocalGurdianScema = new Schema<LocalGurdian>({
  name: { type: String, required: true },
  occupation: { type: String, required: true },
  contactNo: { type: String, required: true },
  address: { type: String, required: true },
});

const StudentScema = new Schema<Student>({
  id: { type: String },
  name: UserNameScema,
  gender: ['female', 'male'],
  dateOfBirth: { type: String },
  email: { type: String, required: true },
  contactNo: { type: String, required: true },
  emergencyContactNo: { type: String, required: true },
  bloodGroup: ['A+', 'A-', 'B+', 'B-', 'AB+', 'AB-', 'O+', 'O-'],
  presentAddress: { type: String, required: true },
  permanentAddress: { type: String, required: true },
  guardian: GuardianScema,
  localGuardian: LocalGurdianScema,
  profileImage: { type: String, required: true },
  isActive: ['active', 'inActive'],
});

export const StudentModel = model<Student>('Student', StudentScema);
